Change dividend ID column type from BigInteger to Integer for consistency in the DividendHistory model

+++ backend/app/models/dividend.py
@@ -1,11 +1,11 @@
 from app.models.base import Base
-from sqlalchemy import BigInteger, Column, Date, Float, ForeignKey, String, UniqueConstraint
+from sqlalchemy import Column, Date, Float, ForeignKey, Integer, String, UniqueConstraint
 
 
 class DividendHistory(Base):
     __tablename__ = "dividend_history"
 
-    id = Column(BigInteger, primary_key=True, autoincrement=True)
+    id = Column(Integer, primary_key=True, autoincrement=True)
     ticker = Column(String, ForeignKey("assets.ticker"), index=True, nullable=False)
     date = Column(Date, index=True, nullable=False)
